Added grey spritesheets.

+++ Sonic/Sonic.Server/entities/Server.cpp
@@ -126,7 +126,7 @@ bool Server::startListening()
 
 int Server::getAvailableIndex()
 {
-	for (size_t i = 0; i < this->serverConfig->getMaxAllowedClients(); i++)
+	for (int i = 0; i < this->serverConfig->getMaxAllowedClients(); i++)
 	{
 		if (!clients.count(i)) return i;
 	}
